teste enviando registro e login
